requirements.txt
pyTelegramBotAPI
requests
translators
beautifulsoup4 
python-dotenv   
urllib3==1.26.15